requirements.txt
Flask>=3.0.0
Flask-Babel>=4.0.0
WordCloud>=1.9.0
nltk>=3.8.1
matplotlib>=3.8.0
gunicorn>=21.2.0
pytest>=7.4.0
python-dotenv>=1.0.0
pylint>=3.0.0
mypy>=1.6.0
black>=23.9.1
